requirements.txt
# Discord Bot Dependencies
discord.py==2.3.2
python-dotenv==1.0.0

# Backend Dependencies
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.5.2

# Shared Dependencies
pymongo==4.6.1
python-dateutil==2.8.2
requests==2.31.0

# Optional Dependencies
python-jose==3.3.0  # For JWT tokens
passlib==1.7.4  # For password hashing 